requirements.txt
Django==3.1
Pillow==7.1
certifi==2020.6.20
chardet==3.0.4
cssselect==1.1.0
html2text==2020.1.16
idna==2.10
lxml==4.5.2
readability-lxml==0.8.1
requests==2.24.0
soupsieve==2.0.1
functions-framework==2.0.0
google-cloud-firestore==1.9.0